- implemented a new hook `useAnalytics` as an alternative to the `AnalyticsComponent` - extracted components and classes from the analytics/ folder into separate files
